+++ src/main/java/com/sitanInfo/API_WS_ETUDE/services/GradeService.java
@@ -1,51 +1,40 @@
 package com.sitanInfo.API_WS_ETUDE.services;
 
-import com.sitanInfo.API_WS_ETUDE.model.Grades;
+import com.sitanInfo.API_WS_ETUDE.model.Grade;
+import com.sitanInfo.API_WS_ETUDE.model.TypeFormation;
 import com.sitanInfo.API_WS_ETUDE.repository.GradeRepository;
 import lombok.Data;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
-
 @Service
 @Data
-public class GradeService{
+public class GradeService {
 
     @Autowired
     private GradeRepository gradeRepository;
 
-
-    public String creer(Grades grades) {
+    public String creer(Grade grade){
         try {
-            Grades gradeExiste = gradeRepository.getByCode(grades.getCode());
+            Grade gradeExiste = gradeRepository.getByCode(grade.getCode());
             if (gradeExiste != null){
-                return "Cette grade existe";
+                return "Cette grade existe deja";
             } else {
-                gradeRepository.save(grades);
-                return "Grade créer";
+                gradeRepository.save(grade);
+                return "Grade  créer";
             }
-        } catch (Exception e){
+        }catch (Exception e){
             e.printStackTrace();
-            return "Une erreur est survenue lors de la création du grade";
+            return "Une erreur s'est produite lors de la creation";
         }
     }
 
-    public List<Grades> lire() {
-        return gradeRepository.findAll();
-    }
-
-    public Optional<Grades> findByid(Integer id) {
-        return gradeRepository.findById(id);
-    }
-
-    public String supprimer(Integer id) {
-        gradeRepository.deleteById(id);
-        return "Grade non trouvée";
-    }
-
-    public String modifier(Integer id, Grades grades) {
-        return null;
+    public String supprimer(Integer id){
+        if (gradeRepository.existsById(id)){
+            gradeRepository.deleteById(id);
+            return "Grade supprimé";
+        } else {
+            return "Cette grade n'existe pas";
+        }
     }
 }
